Add/remove/switch class(es) with animating.

These functions are the R wrappers of addClass(), removeClass(), switchClass() in jQuery UI library. They should be used in server of a shiny document.

jqui_add_class(selector, className, duration = 400, easing = "swing",
  complete = NULL)

jqui_remove_class(selector, className, duration = 400, easing = "swing",
  complete = NULL)

jqui_switch_class(selector, removeClassName, addClassName, duration = 400,
  easing = "swing", complete = NULL)

Arguments

selector

A jQuery's selector that determines the shiny tag element(s) whose interaction is enable or disable interactions.

className

One or more class names (space separated) to be added to or removed from the class attribute of each matched element.

duration

A string or number determining how long the animation will run.

easing

A string indicating which easing function to use for the transition.

complete

A js function to call once the animation is complete, called once per matched element.

removeClassName

One or more class names (space separated) to be removed from the class attribute of each matched element.

addClassName

One or more class names (space separated) to be added to the class attribute of each matched element.

Details

jqui_add_class

Adds the specified class(es) to each of the set of matched elements while animating all style changes.

jqui_remove_class

Removes the specified class(es) from each of the set of matched elements while animating all style changes.

jqui_switch_class

Adds and removes the specified class(es) to each of the set of matched elements while animating all style changes

Examples

## Not run: ------------------------------------
#   # in shiny ui create a span
#   tags$span(id = 'foo', 'class animation demo')
# 
#   # in shiny server add class 'lead' to the span
#   jqui_add_class('#foo', className = 'lead')
## ---------------------------------------------