Create a draggable modal dialog UI

This creates the UI for a modal dialog similar to shiny::modalDialog except its content is draggable.

draggableModalDialog(
  ...,
  title = NULL,
  footer = shiny::modalButton("Dismiss"),
  size = c("m", "s", "l"),
  easyClose = FALSE,
  fade = TRUE
)

Arguments

...
UI elements for the body of the modal dialog box.

title
An optional title for the dialog.

footer
UI for footer. Use NULL for no footer.

size
One of "s" for small, "m" (the default) for medium, or "l" for large.

easyClose
If TRUE, the modal dialog can be dismissed by clicking outside the dialog box, or be pressing the Escape key. If FALSE (the default), the modal dialog can't be dismissed in those ways; instead it must be dismissed by clicking on the dismiss button, or from a call to removeModal() on the server.

fade
If FALSE, the modal dialog will have no fade-in animation (it will simply appear rather than fade in to view).

Value

A modified shiny modal dialog UI with its content draggable.