A shiny input control to show the order of a list of items

Display a list items whose order can be changed by drag and drop within or between orderInput(s). The current items order can be obtained from input$inputId_order.

orderInput(
  inputId,
  label,
  items,
  as_source = FALSE,
  connect = NULL,
  item_class = c("default", "primary", "success", "info", "warning", "danger"),
  placeholder = NULL,
  width = "500px",
  ...
)

Arguments

inputId
The input slot that will be used to access the current order of items.

label
Display label for the control, or NULL for no label.

items
Items to display, can be a list, an atomic vector or a factor. For list and atomic vector, if named, the names are displayed and the order is given in values. For factor, values are displayed and the order is given in levels

as_source
A boolean value to determine whether the orderInput is set as source mode. If source mode, items in this orderInput can only be dragged (copied) to the connected non-source orderInput(s) defined by connect argument. If non-source mode, items in the orderInput can be dragged (moved) within or toward other connected non-source orderInput(s) defined by connect argument.

connect
optional, a vector of inputId(s) of other orderInput(s) that the current orderInput connects to. The behavior of the connected orderInput(s) depend on the as_source argument.

item_class
One of the Bootstrap Button Styles to apply to each item.

placeholder
A character string to show when there is no item left in the orderInput.

width
The width of the input, e.g. '400px', or '100\ validateCssUnit.

...
Arguments passed to shiny::tags$div which is used to build the container of the orderInput.

Value

A orderInput control that can be added to a UI definition.

Examples

orderInput('items1', 'Items1', items = month.abb, item_class = 'info')

#> <label for="items1">Items1</label>
#> <div id="items1" style="width: 500px; font-size: 0px; min-height: 25px;">
#>   <div data-value="Jan" class="btn btn-info" style="margin: 1px">Jan</div>
#>   <div data-value="Feb" class="btn btn-info" style="margin: 1px">Feb</div>
#>   <div data-value="Mar" class="btn btn-info" style="margin: 1px">Mar</div>
#>   <div data-value="Apr" class="btn btn-info" style="margin: 1px">Apr</div>
#>   <div data-value="May" class="btn btn-info" style="margin: 1px">May</div>
#>   <div data-value="Jun" class="btn btn-info" style="margin: 1px">Jun</div>
#>   <div data-value="Jul" class="btn btn-info" style="margin: 1px">Jul</div>
#>   <div data-value="Aug" class="btn btn-info" style="margin: 1px">Aug</div>
#>   <div data-value="Sep" class="btn btn-info" style="margin: 1px">Sep</div>
#>   <div data-value="Oct" class="btn btn-info" style="margin: 1px">Oct</div>
#>   <div data-value="Nov" class="btn btn-info" style="margin: 1px">Nov</div>
#>   <div data-value="Dec" class="btn btn-info" style="margin: 1px">Dec</div>
#> </div>
#> <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":"[id='items1']","type":"interaction","func":"sortable","operation":"enable","options":{"connectWith":"false","remove":"function(e, ui){if(!$(e.target).children().length)$(e.target).empty();}","shiny":{"order":{"sortcreate":"function(event, ui){\n      return $(event.target).children().map(function(i, e){\n        return $(e).attr(\"data-value\");\n      }).get();\n    }","sortupdate":"function(event, ui){\n      return $(event.target).children().map(function(i, e){\n        return $(e).attr(\"data-value\");\n      }).get();\n    }"}}},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectWith":false,"remove":true,"shiny":{"order":{"sortcreate":true,"sortupdate":true}}},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>

## build connections between orderInputs
orderInput('items2', 'Items2 (can be moved to Items1 and Items4)', items = month.abb,
           connect = c('items1', 'items4'), item_class = 'primary')

#> <label for="items2">Items2 (can be moved to Items1 and Items4)</label>
#> <div id="items2" style="width: 500px; font-size: 0px; min-height: 25px;">
#>   <div data-value="Jan" class="btn btn-primary" style="margin: 1px">Jan</div>
#>   <div data-value="Feb" class="btn btn-primary" style="margin: 1px">Feb</div>
#>   <div data-value="Mar" class="btn btn-primary" style="margin: 1px">Mar</div>
#>   <div data-value="Apr" class="btn btn-primary" style="margin: 1px">Apr</div>
#>   <div data-value="May" class="btn btn-primary" style="margin: 1px">May</div>
#>   <div data-value="Jun" class="btn btn-primary" style="margin: 1px">Jun</div>
#>   <div data-value="Jul" class="btn btn-primary" style="margin: 1px">Jul</div>
#>   <div data-value="Aug" class="btn btn-primary" style="margin: 1px">Aug</div>
#>   <div data-value="Sep" class="btn btn-primary" style="margin: 1px">Sep</div>
#>   <div data-value="Oct" class="btn btn-primary" style="margin: 1px">Oct</div>
#>   <div data-value="Nov" class="btn btn-primary" style="margin: 1px">Nov</div>
#>   <div data-value="Dec" class="btn btn-primary" style="margin: 1px">Dec</div>
#> </div>
#> <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":"[id='items2']","type":"interaction","func":"sortable","operation":"enable","options":{"connectWith":"#items1, #items4","remove":"function(e, ui){if(!$(e.target).children().length)$(e.target).empty();}","shiny":{"order":{"sortcreate":"function(event, ui){\n      return $(event.target).children().map(function(i, e){\n        return $(e).attr(\"data-value\");\n      }).get();\n    }","sortupdate":"function(event, ui){\n      return $(event.target).children().map(function(i, e){\n        return $(e).attr(\"data-value\");\n      }).get();\n    }"}}},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectWith":false,"remove":true,"shiny":{"order":{"sortcreate":true,"sortupdate":true}}},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>

## build connections in source mode
orderInput('items3', 'Items3 (can be copied to Items2 and Items4)', items = month.abb,
           as_source = TRUE, connect = c('items2', 'items4'), item_class = 'success')

#> <label for="items3">Items3 (can be copied to Items2 and Items4)</label>
#> <div id="items3" style="width: 500px; font-size: 0px; min-height: 25px;">
#>   <div class="btn btn-success jqui-interaction-SwlKLUUE" data-value="Jan" style="margin: 1px">Jan</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-SwlKLUUE","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-dLieexom" data-value="Feb" style="margin: 1px">Feb</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-dLieexom","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-b92Q9Qeu" data-value="Mar" style="margin: 1px">Mar</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-b92Q9Qeu","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-R8HfdVHI" data-value="Apr" style="margin: 1px">Apr</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-R8HfdVHI","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-yvQgF0fi" data-value="May" style="margin: 1px">May</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-yvQgF0fi","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-vjF0HqZ3" data-value="Jun" style="margin: 1px">Jun</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-vjF0HqZ3","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-vEafHYTa" data-value="Jul" style="margin: 1px">Jul</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-vEafHYTa","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-yUXKhGFR" data-value="Aug" style="margin: 1px">Aug</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-yUXKhGFR","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-sBBVkzIb" data-value="Sep" style="margin: 1px">Sep</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-sBBVkzIb","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-wPcLdPaM" data-value="Oct" style="margin: 1px">Oct</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-wPcLdPaM","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-fvs4bNaB" data-value="Nov" style="margin: 1px">Nov</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-fvs4bNaB","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#>   <div class="btn btn-success jqui-interaction-Ggl1OPbL" data-value="Dec" style="margin: 1px">Dec</div>
#>   <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":".jqui-interaction-Ggl1OPbL","type":"interaction","func":"draggable","operation":"enable","options":{"connectToSortable":"#items2, #items4","helper":"clone","cancel":""},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectToSortable":false,"helper":false,"cancel":false},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>
#> </div>

## show placeholder
orderInput('items4', 'Items4 (can be moved to Items2)', items = NULL, connect = 'items2',
           placeholder = 'Drag items here...')

#> <label for="items4">Items4 (can be moved to Items2)</label>
#> <div id="items4" style="width: 500px; font-size: 0px; min-height: 25px;"></div>
#> <script class="jqui_self_cleaning_script">$(function(){setTimeout(function(){shinyjqui.msgCallback({"ui":"[id='items4']","type":"interaction","func":"sortable","operation":"enable","options":{"connectWith":"#items2","remove":"function(e, ui){if(!$(e.target).children().length)$(e.target).empty();}","shiny":{"order":{"sortcreate":"function(event, ui){\n      return $(event.target).children().map(function(i, e){\n        return $(e).attr(\"data-value\");\n      }).get();\n    }","sortupdate":"function(event, ui){\n      return $(event.target).children().map(function(i, e){\n        return $(e).attr(\"data-value\");\n      }).get();\n    }"}}},"debug":false,"_js_idx":{"ui":false,"type":false,"func":false,"operation":false,"options":{"connectWith":false,"remove":true,"shiny":{"order":{"sortcreate":true,"sortupdate":true}}},"debug":false}});
#>       $(".jqui_self_cleaning_script").remove();}, 10);});</script>