Initialize, enable or disable jQuery UI interactions

These functions initialize or switch on/off jQuery UI mouse interactions (draggable, droppable, resizable, selectable, sortable) to shiny tag element(s).

jqui_resizabled(tag, options = NULL)

jqui_sortabled(tag, options = NULL)

jqui_draggabled(tag, options = NULL)

jqui_droppabled(tag, options = NULL)

jqui_selectabled(tag, options = NULL)

jqui_sortable(selector, switch = TRUE, options = NULL)

jqui_draggable(selector, switch = TRUE, options = NULL)

jqui_droppable(selector, switch = TRUE, options = NULL)

jqui_selectable(selector, switch = TRUE, options = NULL)

jqui_resizable(selector, switch = TRUE, options = NULL)

Arguments

tag

A shiny tag object to add interactions to.

options

A list of interaction options. Ignored when switch is set as FALSE. In addition to the interaction_specific_options, this parameter also accept a shiny option that controls the shiny input value returned from the element. See Details.

selector

A jQuery_selector that determines the shiny tag element(s) whose interaction is to be enabled or disabled.

switch

A boolean value to determine whether to enable or disable an interaction.

Value

jqui_draggabled, jqui_droppabled, jqui_resizabled, jqui_selectabled and jqui_sortabled returns a modified shiny tag object with interaction enabled.

Details

The function jqui_draggabled, jqui_droppabled, jqui_resizabled, jqui_selectabled and jqui_sortabled initialize the element's interactions and should be used in ui of a shiny app. The function jqui_draggable, jqui_droppable, jqui_resizable, jqui_selectable and jqui_sortable switch on/off interactions and should be used in server of a shiny app.

If an element has an id and its interaction is initialized or switched on, users can get access to some internally defined shiny input values. Users can override the default shiny input settings by passing a shiny option to the options parameter. Please see the vignette Introduction to shinyjqui for more details.

Examples

library(shiny)
library(highcharter)

## used in ui
jqui_resizabled(actionButton('btn', 'Button'))
jqui_draggabled(plotOutput('plot', width = '400px', height = '400px'),
                options = list(axis = 'x'))
jqui_selectabled(
  div(
    id = 'sel_plots',
    highchartOutput('highchart', width = '300px'),
    plotOutput('ggplot', width = '300px')
  ),
  options = list(
    classes = list(`ui-selected` = 'ui-state-highlight')
  )
)
jqui_sortabled(tags$ul(
  id = 'lst',
  tags$li('A'),
  tags$li('B'),
  tags$li('C')
))

## used in server
## Not run: ------------------------------------
#   jqui_draggable('#foo', options = list(grid = c(80, 80)))
#   jqui_droppable('.foo', switch = FALSE)
## ---------------------------------------------

## use shiny input
if (interactive()) {
  server <- function(input, output) {
    output$foo <- renderHighchart({
      hchart(mtcars, "scatter", hcaes(x = cyl, y = mpg))
    })
    output$position <- renderPrint({
      print(input$foo_position)
    })
  }
  ui <- fluidPage(
    includeJqueryUI(),
    verbatimTextOutput('position'),
    jqui_draggabled(highchartOutput('foo', width = '200px', height = '200px'))
  )
  shinyApp(ui, server)
}

## custom shiny input
func <- htmlwidgets::JS('function(event, ui){return $(event.target).offset();}')
options <-  list(
  shiny = list(
    abs_position = list(
      dragcreate = func, # send returned value back to shiny when interaction is created.
      drag = func # send returned value to shiny when dragging.
    )
  )
)
jqui_draggabled(highchartOutput('foo', width = '200px', height = '200px'),
                options = options)